chore: renamed Game type to Match and updated its props
